bugfix:修复节点tooltip显示不正确的问题 #869
