+++ frontend/mobile/src/components/MobileCanvas/MobileNodeTemplate.vue
@@ -24,7 +24,7 @@
                 :class="['bk-flow-location', node['status'] ? node['status'].toLowerCase() : '']"
                 @click="onNodeClick(node, $event)">
                 <div class="node-name">
-                    <p class="name">{{ node.name }}</p>
+                    <p class="name" :data-id="node.id">{{ node.name }}</p>
                 </div>
                 <div class="task-name">{{ node.stage_name }}</div>
             </div>
@@ -52,11 +52,6 @@
 </template>
 
 <script>
-    import Tooltip from 'tooltip.js'
-    import { mapActions, mapState } from 'vuex'
-    import { errorHandler } from '@/utils/errorHandler.js'
-
-    const NODE_ACTION = ['retry', 'skip', 'detail', 'timer', 'resume']
 
     export default {
         name: 'MobileNodeTemplate',
@@ -68,49 +63,15 @@
                 }
             }
         },
-        inject: ['refreshTaskStatus'],
         data () {
             return {
-                nodeTooltipInstance: {},
                 i18n: {
                     start: window.gettext('开始'),
-                    end: window.gettext('结束'),
-                    detail: window.gettext('执行详情'),
-                    retry: window.gettext('重试'),
-                    skip: window.gettext('跳过'),
-                    sub: window.gettext('查看子流程'),
-                    editTime: window.gettext('修改时间'),
-                    skipSuccess: window.gettext('跳过成功'),
-                    skipFailed: window.gettext('跳过失败'),
-                    resume: window.gettext('继续')
+                    end: window.gettext('结束')
                 }
             }
         },
-        computed: {
-            ...mapState({
-                taskId: state => state.taskId
-            })
-        },
-        destroyed () {
-            this.$el.removeEventListener('click', this.handleNodeClick, false)
-            this.clearNodeTooltipInstance()
-        },
-        mounted () {
-            this.$el.addEventListener('click', this.handleNodeClick, false)
-        },
         methods: {
-            ...mapActions('task', [
-                'getTask',
-                'instanceNodeResume',
-                'instanceNodeSkip',
-                'getNodeRetryData'
-            ]),
-            handleNodeClick (e) {
-                const nodeAction = global.$(e.target).data('action')
-                if (nodeAction) {
-                    this.onNodeOperationClick(nodeAction)
-                }
-            },
             onNodeClick (node) {
                 const pipelineTree = this.$store.state.pipelineTree
                 if (this.node.type === 'subflow') {
@@ -119,154 +80,7 @@
                         // 当前树替换成子流程树
                         this.$store.commit('setPipelineTree', subTree)
                     }
-                } else {
-                    const componentCode = pipelineTree.activities[node.id].component.code
-                    if (!this.node.componentCode) {
-                        this.$set(this.node, 'componentCode', componentCode)
-                    }
-                    if (this.node.type === 'tasknode' && this.node.status) {
-                        let tip = null
-                        if (!this.nodeTooltipInstance[node.id]) {
-                            const el = global.$(`[name=tip_${node.id}]`)[0]
-                            tip = new Tooltip(el, {
-                                placement: 'bottom',
-                                html: true,
-                                title: this.getNodeBtnTpl(node)
-                            })
-                            this.nodeTooltipInstance[node.id] = tip
-                        } else {
-                            tip = this.nodeTooltipInstance[node.id]
-                            tip.hide()
-                        }
-                        tip.show()
-                    }
                 }
-            },
-
-            getNodeBtnTpl (node) {
-                const btnList = []
-                if (node.status === 'RUNNING') {
-                    if (node.componentCode === 'sleep_timer') {
-                        btnList.push({ type: 'timer', text: this.i18n.editTime })
-                    }
-                    btnList.push({ type: 'resume', text: this.i18n.resume })
-                } else if (node.status === 'FAILED') {
-                    btnList.push({ type: 'retry', text: this.i18n.retry })
-                    btnList.push({ type: 'skip', text: this.i18n.skip })
-                }
-                btnList.push({ type: 'detail', text: this.i18n.detail })
-                return this.nodeBtnTplFactory(btnList)
-            },
-
-            nodeBtnTplFactory (btnList) {
-                return `
-                    <div class="btn-wrapper">
-                    ${btnList.map(btn => `
-                        <div class="tooltip-btn" data-action="${btn.type}">${btn.text}</div>
-                    `).join('')}
-                    </div>
-                    `
-            },
-
-            onNodeOperationClick (operation) {
-                if (!this.operating) {
-                    this.operating = true
-                    if (NODE_ACTION.includes(operation)) {
-                        this[operation]()
-                    }
-                    this.operating = false
-                }
-            },
-            detail () {
-                this.$store.commit('setNode', this.node)
-                this.$router.push({ name: 'task_nodes' })
-            },
-            destroyTooltipInstance (id) {
-                if (this.nodeTooltipInstance[id]) {
-                    this.nodeTooltipInstance[id].dispose()
-                    delete this.nodeTooltipInstance[id]
-                }
-            },
-            clearNodeTooltipInstance () {
-                Object.keys(this.nodeTooltipInstance).forEach(item => {
-                    this.destroyTooltipInstance(item)
-                })
-            },
-            async retry () {
-                this.$toast.loading({ mask: true, message: this.i18n.loading })
-                const task = await this.getTask({ id: this.taskId })
-                const pipelineTree = JSON.parse(task.pipeline_tree)
-                const taskId = this.taskId
-                const params = {
-                    taskId: taskId,
-                    nodeId: this.node.id,
-                    componentCode: pipelineTree.activities[this.node.id].component.code
-                }
-                try {
-                    const response = await this.getNodeRetryData(params)
-                    params.inputs = response.data.inputs
-                    const newInputs = {}
-                    for (const k of Object.keys(params.inputs)) {
-                        newInputs[`${params.componentCode}.${k}`] = { source_tag: `${params.componentCode}.${k}`, value: params.inputs[k] }
-                    }
-                    params.inputs = newInputs
-                } catch (e) {
-                    errorHandler(e, this)
-                }
-                this.$toast.clear()
-                this.operating = false
-                this.$router.push({ name: 'task_reset', params: params })
-            },
-            async skip () {
-                try {
-                    this.$toast.loading({ mask: true, message: this.i18n.loading })
-                    const response = await this.instanceNodeSkip({ id: this.taskId, nodeId: this.node.id })
-                    if (response.result) {
-                        global.bus.$emit('notify', { message: this.i18n.skipSuccess })
-                        setTimeout(() => {
-                            this.refreshTaskStatus()
-                        }, 1000)
-                    } else {
-                        errorHandler(response, this)
-                    }
-                } catch (e) {
-                    errorHandler(e, this)
-                } finally {
-                    this.$toast.clear()
-                    this.operating = false
-                    this.clearNodeTooltipInstance()
-                }
-            },
-            async resume () {
-                try {
-                    const response = await this.instanceNodeResume({ id: this.taskId, nodeId: this.node.id })
-                    if (response.result) {
-                        this.refreshTaskStatus()
-                    } else {
-                        errorHandler(response, this)
-                    }
-                } catch (e) {
-                    errorHandler(e, this)
-                } finally {
-                    this.operating = false
-                    this.clearNodeTooltipInstance()
-                }
-            },
-            async timer () {
-                this.$toast.loading({ mask: true, message: this.i18n.loading })
-                const task = await this.getTask({ id: this.taskId })
-                const pipelineTree = JSON.parse(task.pipeline_tree)
-                const taskId = this.taskId
-                const params = {
-                    taskId: taskId,
-                    nodeId: this.node.id,
-                    componentCode: pipelineTree.activities[this.node.id].component.code
-                }
-                const response = await this.getNodeRetryData(params)
-                params.inputs = response.data.inputs
-                this.$toast.clear()
-                this.operating = false
-                this.$router.push({ name: 'task_edit_timing', params: params })
             }
         }
     }
